functions/pdf-vision-pipeline/requirements.txt
functions-framework==3.5.0
google-cloud-vision==3.4.4
google-cloud-secret-manager==2.17.0
pdf2image==1.17.0
Pillow==10.1.0
boto3==1.34.0
botocore>=1.34.0
requests==2.31.0 